requirements.txt
streamlit==1.25.0
pandas==2.1.0
qrcode==7.4.2
Pillow==10.0.1